[#9708] move getAllAfter up
