+++ sormas-api/src/main/java/de/symeda/sormas/api/BaseFacade.java
@@ -29,6 +29,8 @@ public interface BaseFacade<DTO extends EntityDto, INDEX_DTO extends Serializabl
 
 	List<DTO> getAllAfter(Date date);
 
+	List<DTO> getAllAfter(Date date, Integer batchSize, String lastSynchronizedUuid);
+
 	List<String> getObsoleteUuidsSince(Date since);
 
 	void validate(@Valid DTO dto) throws ValidationRuntimeException;
