+++ sormas-api/src/main/java/de/symeda/sormas/api/person/PersonFacade.java
@@ -46,8 +46,6 @@ public interface PersonFacade extends BaseFacade<PersonDto, PersonIndexDto, Pers
 
 	DataHelper.Pair<CaseClassification, PersonDto> savePersonWithoutNotifyingExternalJournal(@Valid PersonDto source);
 
-	List<PersonDto> getAllAfter(Date date, Integer batchSize, String lastSynchronizedUuid);
-
 	/**
 	 * Returns a list with the names of all persons that the user has access to and that match the criteria.
 	 * This only includes persons that are associated with an active case, contact or event participant.
